+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 0.5.0 (Next)
 
 * `SlackRubyBot::Server` supports `restart!` with retry - [@dblock](https://github.com/dblock).
-* `SlackRubyBot::Server` publicly supports `auth!` and `start!` that make up a `run` loop - [@dblock](https://github.com/dblock).
+* `SlackRubyBot::Server` publicly supports `auth!`, `start!` and `start_async` that make up a `run` loop - [@dblock](https://github.com/dblock).
 * Extracted `SlackRubyBot::Server` from `SlackRubyBot::App` - [@dblock](https://github.com/dblock).
 * Fix: explicitly require 'giphy' - [@dblock](https://github.com/dblock).
 * Fix: undefined method `stop` for `Slack::RealTime::Client` - [@dblock](https://github.com/dblock).